Adding service network adapter
